feat: enable installation as cli
